More caching and more logging
I've added some more logging on the server side (in RZC) that times
different portions of the task.

This looks like:
  Rzr Summary:
  CreateEngine:        00:00:00.0213964
  TagHelpers:          00:00:00.0882738
  GenerateCSharp:      00:00:00.0411723
  CreateCompilation:   00:00:00.0062635
  GenerateCSharp:      00:00:00.0000010
  EmitAssembly:        00:00:00.0655476
  Total:               00:00:00.2227173

Compare this with the numbers returned by the MSBuild performance
summary, which is more like 793ms at this time. So there's a 500ms perf
gap to resolve here by looking at what the client is doing.